requirements.txt
pandas
gradio
google.generativeai
llama_index.embeddings.huggingface
sentence-transformers